Allow disabling address resolution
When client does DNS lookup by itself and provides raw IP addresses
to XMemcachedClient - it should be possible to disable any automatic
hostname resolution. Otherwise it might cause conflicts.

For the same purpose client should be able to remove a server using
raw address, not a host name. Exposed method should be renamed to
removeServer to be consistent with a complimentary pair of addServer
methods.